requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
future==1.0.0
geocoder==1.38.1
idna==3.7
packaging==24.0
pillow==10.3.0
pyperclip==1.8.2
ratelim==0.1.6
requests==2.32.3
six==1.16.0
tkintermapview==1.29
urllib3==2.2.1
